review: address lavra-review findings
- Deduplicate bearer auth into src/auth.rs (was triplicated across otlp.rs / mcp/routes.rs / api.rs — security code drift hazard)
- scrub_secrets uses Cow<str> threading: zero allocation when no pattern matches (80-95% allocator reduction on AI burst path)
- PEM private key block scrubber matches whole block (was: BEGIN line only, leaving key body verbatim and FTS5-indexed)
- Add standalone JWT regex pattern; broaden Authorization Bearer charset to include +/=
- password=/api_key= matcher handles quoted values
- Merge severity_from_number 1..=4 and 5..=8 into 1..=8 → "debug" (both OTLP TRACE and DEBUG map there)
- Startup tracing::warn! when OTLP runs without auth on a non-loopback bind

187 lib tests pass, clippy clean.
