+++ src/mcp/routes.rs
@@ -10,11 +10,12 @@ use axum::{
     Router,
 };
 use serde_json::json;
-use subtle::ConstantTimeEq;
 use tower_http::{
     cors::{Any, CorsLayer},
     limit::RequestBodyLimitLayer,
 };
+
+use crate::auth::{bearer_token, token_matches};
 
 use super::rmcp_server::allowed_origins;
 use super::AppState;
@@ -101,35 +102,6 @@ fn cors_layer(config: &crate::config::McpConfig) -> CorsLayer {
         .allow_origin(origins)
         .allow_methods([Method::POST, Method::GET])
         .allow_headers(Any)
-}
-
-fn bearer_token(auth: &str) -> Option<&str> {
-    let mut parts = auth.split_whitespace();
-    let scheme = parts.next()?;
-    let token = parts.next()?;
-    if parts.next().is_some() || !scheme.eq_ignore_ascii_case("bearer") {
-        return None;
-    }
-    Some(token)
-}
-
-fn token_matches(provided: &str, expected: &str) -> bool {
-    const MAX_TOKEN_LEN: usize = 4096;
-    if provided.len() > MAX_TOKEN_LEN || expected.len() > MAX_TOKEN_LEN {
-        return false;
-    }
-
-    let mut provided_buf = [0_u8; MAX_TOKEN_LEN];
-    let mut expected_buf = [0_u8; MAX_TOKEN_LEN];
-    provided_buf[..provided.len()].copy_from_slice(provided.as_bytes());
-    expected_buf[..expected.len()].copy_from_slice(expected.as_bytes());
-
-    let bytes_match = provided_buf.ct_eq(&expected_buf).unwrap_u8() == 1;
-    let lengths_match = (provided.len() as u64)
-        .ct_eq(&(expected.len() as u64))
-        .unwrap_u8()
-        == 1;
-    bytes_match && lengths_match
 }
 
 /// Health check — lightweight probe that verifies DB connectivity without
